requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.30.6
pyswisseph==2.10.3.2
python-dateutil==2.9.0.post0
